fix: restore tlon architecture budget headroom
