+++ crates/app/src/channel/registry.rs
@@ -3165,7 +3165,7 @@ const CHANNEL_REGISTRY: &[ChannelRegistryDescriptor] = &[
     ChannelRegistryDescriptor {
         id: "tlon",
         runtime: None,
-        snapshot_builder: Some(build_tlon_snapshots),
+        snapshot_builder: Some(tlon_support::build_tlon_snapshots),
         selection_order: 205,
         selection_label: "urbit ship bot",
         blurb: "Shipped Tlon outbound surface with config-backed Urbit DMs and group sends through a ship-backed poke API; inbound serve support remains planned.",
@@ -3494,93 +3494,6 @@ fn validate_http_url(field: &str, value: &str, issues: &mut Vec<String>) {
 
     let issue = format!("{field} must use http or https, got {scheme}");
     issues.push(issue);
-}
-
-fn normalize_tlon_status_ship(raw: &str) -> Result<String, String> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return Err("ship is empty".to_owned());
-    }
-
-    let ship_body = trimmed.trim_start_matches('~');
-    if ship_body.is_empty() {
-        return Err("ship is empty".to_owned());
-    }
-
-    let has_invalid_character = ship_body.chars().any(|value| {
-        let is_letter = value.is_ascii_alphabetic();
-        let is_separator = value == '-';
-        !is_letter && !is_separator
-    });
-    if has_invalid_character {
-        return Err("ship must contain only letters and `-`".to_owned());
-    }
-
-    let normalized_ship = ship_body.to_ascii_lowercase();
-    let ship = format!("~{normalized_ship}");
-    Ok(ship)
-}
-
-fn normalize_tlon_status_url(raw: &str) -> Result<String, String> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return Err("url is empty".to_owned());
-    }
-
-    let has_scheme = trimmed.contains("://");
-    let candidate = if has_scheme {
-        trimmed.to_owned()
-    } else {
-        format!("https://{trimmed}")
-    };
-
-    let parsed_url = reqwest::Url::parse(candidate.as_str())
-        .map_err(|error| format!("url is invalid: {error}"))?;
-    let scheme = parsed_url.scheme();
-    let is_http = scheme == "http";
-    let is_https = scheme == "https";
-    if !is_http && !is_https {
-        return Err(format!("url must use http or https, got {scheme}"));
-    }
-    if !parsed_url.username().is_empty() || parsed_url.password().is_some() {
-        return Err("url must not include credentials".to_owned());
-    }
-
-    let path = parsed_url.path();
-    let has_non_root_path = path != "/" && !path.is_empty();
-    let has_query = parsed_url.query().is_some();
-    let has_fragment = parsed_url.fragment().is_some();
-    if has_non_root_path || has_query || has_fragment {
-        return Err("url must not include a path, query, or fragment".to_owned());
-    }
-
-    let hostname = parsed_url
-        .host_str()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .ok_or_else(|| "url hostname is invalid".to_owned())?;
-    let normalized_hostname = hostname.to_ascii_lowercase();
-    let normalized_hostname = normalized_hostname.trim_end_matches('.');
-    if normalized_hostname.is_empty() {
-        return Err("url hostname is invalid".to_owned());
-    }
-
-    let port = parsed_url.port();
-    let is_ipv6 = normalized_hostname.contains(':');
-    let host = if let Some(port) = port {
-        if is_ipv6 {
-            format!("[{normalized_hostname}]:{port}")
-        } else {
-            format!("{normalized_hostname}:{port}")
-        }
-    } else if is_ipv6 {
-        format!("[{normalized_hostname}]")
-    } else {
-        normalized_hostname.to_owned()
-    };
-
-    let normalized_url = format!("{scheme}://{host}");
-    Ok(normalized_url)
 }
 
 #[cfg(test)]
@@ -4460,141 +4373,6 @@ fn build_imessage_snapshots(
                     default_account_source,
                 ),
                 Err(error) => build_invalid_imessage_snapshot(
-                    descriptor,
-                    compiled,
-                    configured_account_id.as_str(),
-                    is_default_account,
-                    default_account_source,
-                    error,
-                ),
-            }
-        })
-        .collect()
-}
-
-fn build_tlon_snapshot_for_account(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    resolved: ResolvedTlonChannelConfig,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-) -> ChannelStatusSnapshot {
-    let mut send_issues = Vec::new();
-
-    let ship = resolved.ship();
-    let normalized_ship = ship.as_deref().map(normalize_tlon_status_ship).transpose();
-    if ship.is_none() {
-        send_issues.push("ship is missing".to_owned());
-    }
-    if let Err(error) = normalized_ship.as_ref() {
-        send_issues.push(error.clone());
-    }
-
-    let url = resolved.url();
-    let normalized_url = url.as_deref().map(normalize_tlon_status_url).transpose();
-    if url.is_none() {
-        send_issues.push("url is missing".to_owned());
-    }
-    if let Err(error) = normalized_url.as_ref() {
-        send_issues.push(error.clone());
-    }
-
-    if resolved.code().is_none() {
-        send_issues.push("code is missing".to_owned());
-    }
-
-    let send_operation = if !compiled {
-        unsupported_operation(
-            TLON_SEND_OPERATION,
-            "binary built without feature `channel-tlon`".to_owned(),
-        )
-    } else if !resolved.enabled {
-        disabled_operation(
-            TLON_SEND_OPERATION,
-            "disabled by tlon account configuration".to_owned(),
-        )
-    } else if !send_issues.is_empty() {
-        misconfigured_operation(TLON_SEND_OPERATION, send_issues)
-    } else {
-        ready_operation(TLON_SEND_OPERATION)
-    };
-
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            TLON_SERVE_OPERATION,
-            "binary built without feature `channel-tlon`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            TLON_SERVE_OPERATION,
-            "tlon inbound serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={}", resolved.configured_account_id),
-        format!("configured_account={}", resolved.configured_account_label),
-        format!("account_id={}", resolved.account.id),
-        format!("account={}", resolved.account.label),
-    ];
-    if let Ok(Some(ship)) = normalized_ship.as_ref() {
-        notes.push(format!("ship={ship}"));
-    }
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    let api_base_url = normalized_url.ok().flatten();
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: resolved.configured_account_id.clone(),
-        configured_account_label: resolved.configured_account_label.clone(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: resolved.enabled,
-        api_base_url,
-        notes,
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
-fn build_tlon_snapshots(
-    descriptor: &ChannelRegistryDescriptor,
-    config: &LoongClawConfig,
-    _runtime_dir: &Path,
-    _now_ms: u64,
-) -> Vec<ChannelStatusSnapshot> {
-    let compiled = cfg!(feature = "channel-tlon");
-    let default_selection = config.tlon.default_configured_account_selection();
-    let default_configured_account_id = default_selection.id.clone();
-    let default_account_source = default_selection.source;
-    config
-        .tlon
-        .configured_account_ids()
-        .into_iter()
-        .map(|configured_account_id| {
-            let is_default_account = configured_account_id == default_configured_account_id;
-            match config
-                .tlon
-                .resolve_account(Some(configured_account_id.as_str()))
-            {
-                Ok(resolved) => build_tlon_snapshot_for_account(
-                    descriptor,
-                    compiled,
-                    resolved,
-                    is_default_account,
-                    default_account_source,
-                ),
-                Err(error) => build_invalid_tlon_snapshot(
                     descriptor,
                     compiled,
                     configured_account_id.as_str(),
@@ -7091,63 +6869,6 @@ fn build_invalid_signal_snapshot(
     }
 }
 
-fn build_invalid_tlon_snapshot(
-    descriptor: &ChannelRegistryDescriptor,
-    compiled: bool,
-    configured_account_id: &str,
-    is_default_account: bool,
-    default_account_source: ChannelDefaultAccountSelectionSource,
-    error: String,
-) -> ChannelStatusSnapshot {
-    let send_operation = if !compiled {
-        unsupported_operation(
-            TLON_SEND_OPERATION,
-            "binary built without feature `channel-tlon`".to_owned(),
-        )
-    } else {
-        misconfigured_operation(TLON_SEND_OPERATION, vec![error.clone()])
-    };
-    let serve_operation = if !compiled {
-        unsupported_operation(
-            TLON_SERVE_OPERATION,
-            "binary built without feature `channel-tlon`".to_owned(),
-        )
-    } else {
-        unsupported_operation(
-            TLON_SERVE_OPERATION,
-            "tlon inbound serve runtime is not implemented yet".to_owned(),
-        )
-    };
-
-    let mut notes = vec![
-        format!("configured_account_id={configured_account_id}"),
-        format!("selection_error={error}"),
-    ];
-    if is_default_account {
-        notes.push("default_account=true".to_owned());
-    }
-    notes.push(format!(
-        "default_account_source={}",
-        default_account_source.as_str()
-    ));
-
-    ChannelStatusSnapshot {
-        id: descriptor.id,
-        configured_account_id: configured_account_id.to_owned(),
-        configured_account_label: configured_account_id.to_owned(),
-        is_default_account,
-        default_account_source,
-        label: descriptor.label,
-        aliases: descriptor.aliases.to_vec(),
-        transport: descriptor.transport,
-        compiled,
-        enabled: false,
-        api_base_url: None,
-        notes,
-        operations: vec![send_operation, serve_operation],
-    }
-}
-
 fn build_invalid_irc_snapshot(
     descriptor: &ChannelRegistryDescriptor,
     compiled: bool,
@@ -7598,6 +7319,11 @@ fn now_ms() -> u64 {
         .as_millis() as u64
 }
 
+mod tlon_support;
+
+#[cfg(test)]
+mod hotspot_tests;
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -7613,26 +7339,6 @@ mod tests {
             Some(ChannelPlatform::Telegram)
         );
         assert_eq!(normalize_channel_platform("discord"), None);
-    }
-
-    #[test]
-    fn normalize_channel_catalog_id_maps_runtime_and_stub_aliases() {
-        assert_eq!(normalize_channel_catalog_id("lark"), Some("feishu"));
-        assert_eq!(normalize_channel_catalog_id(" TELEGRAM "), Some("telegram"));
-        assert_eq!(normalize_channel_catalog_id("discord-bot"), Some("discord"));
-        assert_eq!(normalize_channel_catalog_id("slack"), Some("slack"));
-        assert_eq!(normalize_channel_catalog_id("gchat"), Some("google-chat"));
-        assert_eq!(
-            normalize_channel_catalog_id("synochat"),
-            Some("synology-chat")
-        );
-        assert_eq!(
-            normalize_channel_catalog_id("bluebubbles"),
-            Some("imessage")
-        );
-        assert_eq!(normalize_channel_catalog_id("urbit"), Some("tlon"));
-        assert_eq!(normalize_channel_catalog_id("web-ui"), Some("webchat"));
-        assert_eq!(normalize_channel_catalog_id("unknown"), None);
     }
 
     #[test]
@@ -7689,92 +7395,6 @@ mod tests {
         assert_eq!(
             resolve_channel_runtime_command_descriptor("slack-bot"),
             None
-        );
-    }
-
-    #[test]
-    fn resolve_channel_catalog_command_family_descriptor_includes_runtime_and_stub_channels() {
-        let telegram = resolve_channel_catalog_command_family_descriptor("telegram")
-            .expect("telegram catalog command family");
-        let lark = resolve_channel_catalog_command_family_descriptor("lark")
-            .expect("lark catalog command family");
-        let slack = resolve_channel_catalog_command_family_descriptor("slack-bot")
-            .expect("slack alias catalog command family");
-        let google_chat = resolve_channel_catalog_command_family_descriptor("gchat")
-            .expect("google chat alias catalog command family");
-        let synology_chat = resolve_channel_catalog_command_family_descriptor("synochat")
-            .expect("synology chat alias catalog command family");
-        let irc =
-            resolve_channel_catalog_command_family_descriptor("irc").expect("irc catalog family");
-        let imessage = resolve_channel_catalog_command_family_descriptor("bluebubbles")
-            .expect("imessage alias catalog command family");
-        let tlon = resolve_channel_catalog_command_family_descriptor("urbit")
-            .expect("tlon alias catalog command family");
-
-        assert_eq!(telegram.channel_id, "telegram");
-        assert_eq!(telegram.send.id, CHANNEL_OPERATION_SEND_ID);
-        assert_eq!(telegram.send.command, "telegram-send");
-        assert_eq!(telegram.serve.id, CHANNEL_OPERATION_SERVE_ID);
-        assert_eq!(telegram.serve.command, "telegram-serve");
-        assert_eq!(
-            telegram.default_send_target_kind,
-            ChannelCatalogTargetKind::Conversation
-        );
-
-        assert_eq!(lark.channel_id, "feishu");
-        assert_eq!(lark.send.command, "feishu-send");
-        assert_eq!(lark.serve.command, "feishu-serve");
-        assert_eq!(
-            lark.default_send_target_kind,
-            ChannelCatalogTargetKind::ReceiveId
-        );
-
-        assert_eq!(slack.channel_id, "slack");
-        assert_eq!(slack.send.command, "slack-send");
-        assert_eq!(slack.serve.command, "slack-serve");
-        assert_eq!(
-            slack.default_send_target_kind,
-            ChannelCatalogTargetKind::Conversation
-        );
-
-        assert_eq!(google_chat.channel_id, "google-chat");
-        assert_eq!(google_chat.send.command, "google-chat-send");
-        assert_eq!(google_chat.serve.command, "google-chat-serve");
-        assert_eq!(
-            google_chat.default_send_target_kind,
-            ChannelCatalogTargetKind::Endpoint
-        );
-
-        assert_eq!(synology_chat.channel_id, "synology-chat");
-        assert_eq!(synology_chat.send.command, "synology-chat-send");
-        assert_eq!(synology_chat.serve.command, "synology-chat-serve");
-        assert_eq!(
-            synology_chat.default_send_target_kind,
-            ChannelCatalogTargetKind::Address
-        );
-
-        assert_eq!(irc.channel_id, "irc");
-        assert_eq!(irc.send.command, "irc-send");
-        assert_eq!(irc.serve.command, "irc-serve");
-        assert_eq!(
-            irc.default_send_target_kind,
-            ChannelCatalogTargetKind::Conversation
-        );
-
-        assert_eq!(imessage.channel_id, "imessage");
-        assert_eq!(imessage.send.command, "imessage-send");
-        assert_eq!(imessage.serve.command, "imessage-serve");
-        assert_eq!(
-            imessage.default_send_target_kind,
-            ChannelCatalogTargetKind::Conversation
-        );
-
-        assert_eq!(tlon.channel_id, "tlon");
-        assert_eq!(tlon.send.command, "tlon-send");
-        assert_eq!(tlon.serve.command, "tlon-serve");
-        assert_eq!(
-            tlon.default_send_target_kind,
-            ChannelCatalogTargetKind::Conversation
         );
     }
 
@@ -8177,186 +7797,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn channel_catalog_includes_openclaw_inspired_extended_surfaces() {
-        let catalog = list_channel_catalog();
-        let signal = catalog
-            .iter()
-            .find(|entry| entry.id == "signal")
-            .expect("signal catalog entry");
-        let twitch = catalog
-            .iter()
-            .find(|entry| entry.id == "twitch")
-            .expect("twitch catalog entry");
-        let teams = catalog
-            .iter()
-            .find(|entry| entry.id == "teams")
-            .expect("teams catalog entry");
-        let synology_chat = catalog
-            .iter()
-            .find(|entry| entry.id == "synology-chat")
-            .expect("synology chat catalog entry");
-        let imessage = catalog
-            .iter()
-            .find(|entry| entry.id == "imessage")
-            .expect("imessage catalog entry");
-        let tlon = catalog
-            .iter()
-            .find(|entry| entry.id == "tlon")
-            .expect("tlon catalog entry");
-        let webchat = catalog
-            .iter()
-            .find(|entry| entry.id == "webchat")
-            .expect("webchat catalog entry");
-
-        assert_eq!(
-            signal.supported_target_kinds,
-            vec![ChannelCatalogTargetKind::Address]
-        );
-        assert_eq!(signal.operations[0].command, "signal-send");
-        assert_eq!(signal.operations[1].command, "signal-serve");
-
-        assert_eq!(
-            twitch.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(twitch.selection_order, 135);
-        assert_eq!(twitch.aliases, vec!["tmi"]);
-        assert_eq!(twitch.transport, "twitch_chat_api");
-        assert!(
-            twitch.blurb.contains("Twitch Chat API"),
-            "unexpected twitch blurb: {}",
-            twitch.blurb
-        );
-        assert_eq!(
-            twitch.supported_target_kinds,
-            vec![ChannelCatalogTargetKind::Conversation]
-        );
-        assert_eq!(twitch.operations[0].command, "twitch-send");
-        assert_eq!(twitch.operations[1].command, "twitch-serve");
-        assert_eq!(
-            twitch.operations[0].availability,
-            ChannelCatalogOperationAvailability::Implemented
-        );
-        assert_eq!(
-            twitch.operations[1].availability,
-            ChannelCatalogOperationAvailability::Stub
-        );
-
-        assert_eq!(
-            teams.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(teams.selection_order, 140);
-        assert_eq!(teams.aliases, vec!["msteams", "ms-teams"]);
-        assert_eq!(teams.transport, "microsoft_teams_incoming_webhook");
-        assert_eq!(
-            teams.supported_target_kinds,
-            vec![
-                ChannelCatalogTargetKind::Endpoint,
-                ChannelCatalogTargetKind::Conversation,
-            ]
-        );
-        assert_eq!(teams.operations[0].command, "teams-send");
-        assert_eq!(teams.operations[1].command, "teams-serve");
-        assert_eq!(
-            teams.operations[0].availability,
-            ChannelCatalogOperationAvailability::Implemented
-        );
-        assert_eq!(
-            teams.operations[1].availability,
-            ChannelCatalogOperationAvailability::Stub
-        );
-
-        assert_eq!(
-            synology_chat.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(synology_chat.selection_order, 165);
-        assert_eq!(synology_chat.aliases, vec!["synologychat", "synochat"]);
-        assert_eq!(
-            synology_chat.transport,
-            "synology_chat_outgoing_incoming_webhooks"
-        );
-        assert_eq!(
-            synology_chat.supported_target_kinds,
-            vec![ChannelCatalogTargetKind::Address]
-        );
-        assert_eq!(synology_chat.operations[0].command, "synology-chat-send");
-        assert_eq!(synology_chat.operations[1].command, "synology-chat-serve");
-        assert_eq!(
-            synology_chat.operations[0].availability,
-            ChannelCatalogOperationAvailability::Implemented
-        );
-        assert_eq!(
-            synology_chat.operations[1].availability,
-            ChannelCatalogOperationAvailability::Stub
-        );
-
-        assert_eq!(
-            imessage.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(imessage.aliases, vec!["bluebubbles", "blue-bubbles"]);
-        assert_eq!(imessage.selection_order, 180);
-        assert_eq!(imessage.transport, "imessage_bridge_api");
-        assert!(imessage.blurb.contains("BlueBubbles"));
-        assert_eq!(
-            imessage.supported_target_kinds,
-            vec![ChannelCatalogTargetKind::Conversation]
-        );
-        assert_eq!(imessage.operations[0].command, "imessage-send");
-        assert_eq!(imessage.operations[1].command, "imessage-serve");
-        assert_eq!(
-            imessage.operations[0].availability,
-            ChannelCatalogOperationAvailability::Implemented
-        );
-        assert_eq!(
-            imessage.operations[1].availability,
-            ChannelCatalogOperationAvailability::Stub
-        );
-
-        assert_eq!(
-            tlon.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(tlon.selection_order, 205);
-        assert_eq!(tlon.aliases, vec!["urbit"]);
-        assert_eq!(tlon.transport, "tlon_urbit_ship_api");
-        assert_eq!(
-            tlon.supported_target_kinds,
-            vec![ChannelCatalogTargetKind::Conversation]
-        );
-        assert_eq!(tlon.operations[0].command, "tlon-send");
-        assert_eq!(tlon.operations[1].command, "tlon-serve");
-        assert_eq!(
-            tlon.operations[0].availability,
-            ChannelCatalogOperationAvailability::Implemented
-        );
-        assert_eq!(
-            tlon.operations[1].availability,
-            ChannelCatalogOperationAvailability::Stub
-        );
-        assert_eq!(
-            tlon.onboarding.strategy,
-            ChannelOnboardingStrategy::ManualConfig
-        );
-        assert_eq!(tlon.onboarding.status_command, "loongclaw doctor");
-        assert_eq!(
-            tlon.onboarding.repair_command,
-            Some("loongclaw doctor --fix")
-        );
-
-        assert_eq!(webchat.selection_order, 230);
-        assert_eq!(webchat.aliases, vec!["browser-chat", "web-ui"]);
-        assert_eq!(webchat.transport, "webchat_websocket");
-        assert_eq!(
-            webchat.supported_target_kinds,
-            vec![ChannelCatalogTargetKind::Conversation]
-        );
-    }
-
-    #[test]
     fn channel_catalog_operations_expose_requirement_metadata() {
         let catalog = list_channel_catalog();
         let telegram = catalog
@@ -9108,111 +8548,6 @@ mod tests {
         assert_eq!(webchat.operations[1].command, "webchat-serve");
     }
 
-    #[test]
-    fn channel_inventory_combines_runtime_and_catalog_surfaces() {
-        let config = LoongClawConfig::default();
-        let inventory = channel_inventory(&config);
-
-        assert_eq!(
-            inventory
-                .channels
-                .iter()
-                .map(|snapshot| snapshot.id)
-                .collect::<Vec<_>>(),
-            vec![
-                "telegram",
-                "feishu",
-                "matrix",
-                "wecom",
-                "discord",
-                "slack",
-                "line",
-                "dingtalk",
-                "whatsapp",
-                "email",
-                "webhook",
-                "google-chat",
-                "signal",
-                "twitch",
-                "teams",
-                "mattermost",
-                "nextcloud-talk",
-                "synology-chat",
-                "irc",
-                "imessage",
-                "nostr",
-                "tlon",
-            ]
-        );
-        assert_eq!(
-            inventory
-                .catalog_only_channels
-                .iter()
-                .map(|entry| entry.id)
-                .collect::<Vec<_>>(),
-            vec!["zalo", "zalo-personal", "webchat"]
-        );
-        assert_eq!(
-            inventory
-                .channel_catalog
-                .iter()
-                .map(|entry| entry.id)
-                .collect::<Vec<_>>(),
-            vec![
-                "telegram",
-                "feishu",
-                "matrix",
-                "wecom",
-                "discord",
-                "slack",
-                "line",
-                "dingtalk",
-                "whatsapp",
-                "email",
-                "webhook",
-                "google-chat",
-                "signal",
-                "twitch",
-                "teams",
-                "mattermost",
-                "nextcloud-talk",
-                "synology-chat",
-                "irc",
-                "imessage",
-                "nostr",
-                "tlon",
-                "zalo",
-                "zalo-personal",
-                "webchat",
-            ]
-        );
-        let nostr = inventory
-            .channel_catalog
-            .iter()
-            .find(|entry| entry.id == "nostr")
-            .expect("nostr catalog entry");
-        assert_eq!(
-            nostr.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(
-            nostr.capabilities,
-            vec![ChannelCapability::MultiAccount, ChannelCapability::Send]
-        );
-        assert_eq!(
-            nostr
-                .operations
-                .iter()
-                .map(|operation| operation.availability)
-                .collect::<Vec<_>>(),
-            vec![
-                ChannelCatalogOperationAvailability::Implemented,
-                ChannelCatalogOperationAvailability::Stub,
-            ]
-        );
-    }
-
-    #[test]
     fn channel_catalog_includes_dingtalk_and_google_chat_config_backed_webhook_surfaces() {
         let catalog = list_channel_catalog();
         let dingtalk = catalog
@@ -9508,235 +8843,6 @@ mod tests {
         );
     }
 
-    #[test]
-    fn channel_inventory_exposes_grouped_channel_surfaces() {
-        let mut env = crate::test_support::ScopedEnv::new();
-        env.remove("TELEGRAM_BOT_TOKEN");
-
-        let config = LoongClawConfig::default();
-        let inventory = channel_inventory(&config);
-
-        assert_eq!(
-            inventory
-                .channel_surfaces
-                .iter()
-                .map(|surface| surface.catalog.id)
-                .collect::<Vec<_>>(),
-            vec![
-                "telegram",
-                "feishu",
-                "matrix",
-                "wecom",
-                "discord",
-                "slack",
-                "line",
-                "dingtalk",
-                "whatsapp",
-                "email",
-                "webhook",
-                "google-chat",
-                "signal",
-                "twitch",
-                "teams",
-                "mattermost",
-                "nextcloud-talk",
-                "synology-chat",
-                "irc",
-                "imessage",
-                "nostr",
-                "tlon",
-                "zalo",
-                "zalo-personal",
-                "webchat",
-            ]
-        );
-
-        let telegram = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "telegram")
-            .expect("telegram surface");
-        assert_eq!(telegram.configured_accounts.len(), 1);
-        assert_eq!(
-            telegram.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(telegram.configured_accounts[0].id, "telegram");
-
-        let discord = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "discord")
-            .expect("discord surface");
-        assert_eq!(
-            discord.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(discord.configured_accounts.len(), 1);
-        assert_eq!(
-            discord.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(discord.configured_accounts[0].id, "discord");
-
-        let line = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "line")
-            .expect("line surface");
-        assert_eq!(
-            line.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(line.configured_accounts.len(), 1);
-        assert_eq!(
-            line.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(line.configured_accounts[0].id, "line");
-
-        let wecom = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "wecom")
-            .expect("wecom surface");
-        assert_eq!(
-            wecom.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::RuntimeBacked
-        );
-        assert_eq!(wecom.configured_accounts.len(), 1);
-        assert_eq!(
-            wecom.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(wecom.configured_accounts[0].id, "wecom");
-
-        let mattermost = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "mattermost")
-            .expect("mattermost surface");
-        assert_eq!(
-            mattermost.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(mattermost.configured_accounts.len(), 1);
-        assert_eq!(
-            mattermost.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(mattermost.configured_accounts[0].id, "mattermost");
-
-        let teams = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "teams")
-            .expect("teams surface");
-        assert_eq!(
-            teams.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(teams.configured_accounts.len(), 1);
-        assert_eq!(
-            teams.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(teams.configured_accounts[0].id, "teams");
-
-        let synology_chat = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "synology-chat")
-            .expect("synology chat surface");
-        assert_eq!(
-            synology_chat.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(synology_chat.configured_accounts.len(), 1);
-        assert_eq!(
-            synology_chat.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(synology_chat.configured_accounts[0].id, "synology-chat");
-
-        let imessage = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "imessage")
-            .expect("imessage surface");
-        assert_eq!(
-            imessage.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(imessage.configured_accounts.len(), 1);
-        assert_eq!(
-            imessage.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(imessage.configured_accounts[0].id, "imessage");
-
-        let nostr = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "nostr")
-            .expect("nostr surface");
-        assert_eq!(
-            nostr.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(nostr.configured_accounts.len(), 1);
-        assert_eq!(
-            nostr.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(nostr.configured_accounts[0].id, "nostr");
-
-        let twitch = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "twitch")
-            .expect("twitch surface");
-        assert_eq!(
-            twitch.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(twitch.configured_accounts.len(), 1);
-        assert_eq!(
-            twitch.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(twitch.configured_accounts[0].id, "twitch");
-
-        let tlon = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "tlon")
-            .expect("tlon surface");
-        assert_eq!(
-            tlon.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::ConfigBacked
-        );
-        assert_eq!(tlon.configured_accounts.len(), 1);
-        assert_eq!(
-            tlon.default_configured_account_id.as_deref(),
-            Some("default")
-        );
-        assert_eq!(tlon.configured_accounts[0].id, "tlon");
-
-        let webchat = inventory
-            .channel_surfaces
-            .iter()
-            .find(|surface| surface.catalog.id == "webchat")
-            .expect("webchat surface");
-        assert_eq!(
-            webchat.catalog.implementation_status,
-            ChannelCatalogImplementationStatus::Stub
-        );
-        assert_eq!(webchat.catalog.aliases, vec!["browser-chat", "web-ui"]);
-        assert!(webchat.configured_accounts.is_empty());
-    }
-
-    #[test]
     fn catalog_only_channel_entries_skip_platforms_that_already_have_status_snapshots() {
         let catalog = vec![
             ChannelCatalogEntry {
