+++ crates/daemon/src/lib.rs
@@ -109,11 +109,14 @@ pub mod runtime_restore_cli;
 pub mod skills_cli;
 pub mod source_presentation;
 pub mod supervisor;
+mod tlon_cli;
 
 pub use gateway::read_models::{ChannelsCliJsonPayload, ChannelsCliJsonSchema};
 pub use loongclaw_spec::programmatic::{
     acquire_programmatic_circuit_slot, record_programmatic_circuit_outcome,
 };
+pub use tlon_cli::TLON_SEND_CLI_SPEC;
+use tlon_cli::{default_tlon_send_target_kind, parse_tlon_send_target_kind};
 
 #[allow(
     clippy::expect_used,
@@ -4310,11 +4313,6 @@ pub const TEAMS_SEND_CLI_SPEC: ChannelSendCliSpec = ChannelSendCliSpec {
     run: run_teams_send_cli_impl,
 };
 
-pub const TLON_SEND_CLI_SPEC: ChannelSendCliSpec = ChannelSendCliSpec {
-    family: mvp::channel::TLON_CATALOG_COMMAND_FAMILY_DESCRIPTOR,
-    run: run_tlon_send_cli_impl,
-};
-
 pub const SIGNAL_SEND_CLI_SPEC: ChannelSendCliSpec = ChannelSendCliSpec {
     family: mvp::channel::SIGNAL_CATALOG_COMMAND_FAMILY_DESCRIPTOR,
     run: run_signal_send_cli_impl,
@@ -4593,21 +4591,6 @@ pub fn run_teams_send_cli_impl(args: ChannelSendCliArgs<'_>) -> ChannelCliComman
             args.config_path,
             args.account,
             args.target,
-            args.target_kind,
-            args.text,
-        )
-        .await
-    })
-}
-
-pub fn run_tlon_send_cli_impl(args: ChannelSendCliArgs<'_>) -> ChannelCliCommandFuture<'_> {
-    Box::pin(async move {
-        let _ = args.as_card;
-        let target = require_channel_send_target("tlon-send", args.target)?;
-        mvp::channel::run_tlon_send(
-            args.config_path,
-            args.account,
-            target,
             args.target_kind,
             args.text,
         )
@@ -4906,16 +4889,6 @@ pub fn parse_teams_send_target_kind(
     raw: &str,
 ) -> Result<mvp::channel::ChannelOutboundTargetKind, String> {
     parse_channel_send_target_kind(TEAMS_SEND_CLI_SPEC, raw)
-}
-
-pub fn default_tlon_send_target_kind() -> mvp::channel::ChannelOutboundTargetKind {
-    default_channel_send_target_kind(TLON_SEND_CLI_SPEC)
-}
-
-pub fn parse_tlon_send_target_kind(
-    raw: &str,
-) -> Result<mvp::channel::ChannelOutboundTargetKind, String> {
-    parse_channel_send_target_kind(TLON_SEND_CLI_SPEC, raw)
 }
 
 pub fn default_signal_send_target_kind() -> mvp::channel::ChannelOutboundTargetKind {
